fix datas

+++ ProjectEPI/Data/SectorData.cs
@@ -2,7 +2,8 @@
 {
     public class SectorData
     {
-        public int Id { get; set; } // mudar pra long
+        public long Id { get; set; } // mudar pra long
         public string? Name { get; set; }
+        //public List<EquipmentData>
     }
 }
